Move ApplyDelegate method out of PlatformCrashes and into an iOS extension of Crashes to reduce coupling
