+++ Apps/Contoso.Forms.Puppet/Contoso.Forms.Puppet/ModulePages/CrashesContentPage.xaml.cs
@@ -24,20 +24,7 @@ namespace Contoso.Forms.Puppet
 
         void TestCrash(object sender, System.EventArgs e)
         {
-            try
-            {
-                String[] values = { "one", null, "two" };
-                for (int ctr = 0; ctr <= values.GetUpperBound(0); ctr++)
-                    System.Diagnostics.Debug.WriteLine("{0}{1}", values[ctr].Trim(),
-                                  ctr == values.GetUpperBound(0) ? "" : ", ");
-                System.Diagnostics.Debug.WriteLine("");
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine("null reference exception");
-            }
-
-            //Crashes.GenerateTestCrash();
+            Crashes.GenerateTestCrash();
         }
 
         void DivideByZero(object sender, System.EventArgs e)
@@ -56,6 +43,32 @@ namespace Contoso.Forms.Puppet
         private void GenerateTestCrash(object sender, EventArgs e)
         {
             Crashes.GenerateTestCrash();
+        }
+
+        private void CatchNullReferenceException(object sender, EventArgs e)
+        {
+            try
+            {
+                TriggerNullReferenceException();
+            }
+            catch (NullReferenceException ex)
+            {
+                System.Diagnostics.Debug.WriteLine("null reference exception");
+            }
+        }
+
+        private void CrashWithNullReferenceException(object sender, EventArgs e)
+        {
+            TriggerNullReferenceException();
+        }
+
+        private void TriggerNullReferenceException()
+        {
+            String[] values = { "one", null, "two" };
+            for (int ctr = 0; ctr <= values.GetUpperBound(0); ctr++)
+                System.Diagnostics.Debug.WriteLine("{0}{1}", values[ctr].Trim(),
+                              ctr == values.GetUpperBound(0) ? "" : ", ");
+            System.Diagnostics.Debug.WriteLine("");
         }
 
         private void CrashWithAggregateException(object sender, EventArgs e)
