+++ SDK/MobileCenterCrashes/Microsoft.Azure.Mobile.Crashes/PlatformCrashes.cs
@@ -56,10 +56,6 @@ namespace Microsoft.Azure.Mobile.Crashes
             get { throw new NotImplementedException(); }
         }
 
-        public override void ApplyDelegate()
-        {
-            throw new NotImplementedException();
-        }
         //public override void TrackException(Exception exception)
         //{
         //    throw new NotImplementedException();
